requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
matplotlib===3.9.2
networkx==3.3
openpyxl==3.1.5